Implementation and making tests pass

+++ src/main/java/com/billing/app/domain/billing/application/model/RequestBillItem.java
@@ -1,7 +1,6 @@
 package com.billing.app.domain.billing.application.model;
 
-import lombok.Value;
 import org.springframework.lang.Nullable;
 
-public record RequestBillItem(@Nullable String typeId, String name, int quantity, double price, boolean imported) {
+public record RequestBillItem(@Nullable String typeId, String name, int quantity, int price, boolean imported) {
 }
